register and unregister scripts in the action list

+++ src/database.cpp
@@ -67,9 +67,9 @@ reapack_error Database::lastError() const
   return reapack_error(sqlite3_errmsg(m_db));
 }
 
-uint64_t Database::lastInsertId() const
+int Database::lastInsertId() const
 {
-  return (uint64_t)sqlite3_last_insert_rowid(m_db);
+  return (int)sqlite3_last_insert_rowid(m_db);
 }
 
 int Database::version() const
